matrix view

+++ src/core-app/ducks/middleware/feature/etlView.middleware.js
@@ -17,7 +17,7 @@
 import {
   ETL_VIEW, // feature
   COMPUTE_ETL_VIEW, // command
-  setEtlView, // document
+  setEtlView as setEtlViewAction, // document
   MAKE_DERIVED_FIELD, // command
   addDerivedField as addDerivedFieldAction, // document
   REMOVE_ETL_FIELD, // command
@@ -26,8 +26,8 @@ import {
   resetEtlViewErrors, // document
   deleteDerivedField, // document
 } from '../../actions/etlView.actions';
-import { setHeaderViews } from '../../actions/headerView.actions';
-import { setLoader } from '../../actions/ui.actions';
+import { setHeaderViews as setHeaderViewsAction } from '../../actions/headerView.actions';
+import { setUiLoadingState } from '../../actions/ui.actions';
 import { setNotification } from '../../actions/notifications.actions';
 import {
   addDerivedField,
@@ -43,16 +43,15 @@ import {
   getHeaderViews,
   getEtlFieldChanges,
   isEtlFieldDerived,
+  etlFieldExists,
 } from '../../rootSelectors';
 import { renameEtlField } from '../../../lib/filesToEtlUnits/rename-etl-field';
 import {
   removeNonDerivedEtlField, // triggers async pivot
 } from '../../../lib/filesToEtlUnits/remove-etl-field';
+import { ActionError } from '../../../lib/LuciErrors';
 import { InvalidStateInput } from '../../../lib/LuciFixableErrors';
 import { tagWarehouseState } from '../../actions/workbench.actions';
-import { isHostedWarehouseStateStale } from '../../workbench.reducer';
-
-// import { colors } from '../../../constants/variables';
 
 // -----------------------------------------------------------------------------
 const DEBUG = process.env.REACT_APP_DEBUG_MIDDLEWARE === 'true';
@@ -66,7 +65,7 @@ const DEBUG = process.env.REACT_APP_DEBUG_MIDDLEWARE === 'true';
 const middleware =
   ({ getState }) =>
   (next) =>
-  (action) => {
+  async (action) => {
     //
     if (DEBUG) {
       console.info('👉 loaded etlView.middleware');
@@ -104,23 +103,58 @@ const middleware =
       //      -> workbench
       //
       // -------------------------------------------------------------------------
+      /* 🦀  State mutation error when derived */
       case COMPUTE_ETL_VIEW: {
         //
         // side effect: () -> hvs in state -> etlView
         //
+        const { startTime } = action;
+        next(
+          setUiLoadingState({
+            toggle: true,
+            feature: ETL_VIEW,
+            message: `Computing eltView`,
+          }),
+        );
+        // -------------------
+        //
         const state = getState();
+        const headerViews = getHeaderViews(state);
+        const etlFieldChanges = getEtlFieldChanges(state);
+        // compute the value from state
+        const computedEtlObject = await pivot(headerViews);
 
-        next(setLoader({ toggle: true, feature: ETL_VIEW }));
+        // + user input and other
+        const final = await etlFromPivot()
+          .init(computedEtlObject, etlFieldChanges)
+          .removeStalePropChanges()
+          .removeStaleDerivedFields()
+          .addDerivedFields()
+          .applySourceSequences() // ⚠️  sequence matters (before applyFieldProps)
+          .applyFieldProps()
+          .setGlobalSpanRef()
+          .return();
 
-        const {
-          startTime,
-          headerViews = getHeaderViews(state),
-          etlFieldChanges = getEtlFieldChanges(state),
-        } = action;
-        next(setEtlViewAsync(headerViews, etlFieldChanges, startTime));
+        const endTime = new Date();
+
+        next([
+          setHeaderViewsAction(headerViews),
+          setEtlViewAction(
+            final.etlObject,
+            final.etlFieldChanges,
+            `${endTime - startTime} ms`,
+          ),
+          setUiLoadingState({
+            toggle: false,
+            feature: ETL_VIEW,
+            message: 'Done computing etlView',
+          }),
+        ]);
+
         break;
       }
 
+      /* ✅ All good (no state mutation issues) */
       case MAKE_DERIVED_FIELD: {
         const state = getState();
         const startTime = action?.meta?.startTime || new Date();
@@ -130,7 +164,6 @@ const middleware =
             feature: ETL_VIEW,
           }),
           tagWarehouseState('STALE'),
-
           // field ~ raw input for the field
           // reciever for the normalizing middleware
           // (payload, meta: {normalizer, startTime})
@@ -145,47 +178,65 @@ const middleware =
             startTime,
           }),
         ]);
+
         break;
       }
 
       // pattern: map command -> document
       // action.fieldName
+      // ⬜ try/catch when trying to remove the last etlUnit
       case REMOVE_ETL_FIELD: {
-        // ⬜ try/catch when trying to remove the last etlUnit
+        if (!action?.fieldName) {
+          throw new ActionError(`Remove etl field: missing fieldName`);
+        }
 
-        const state = getState();
+        let state = getState();
+        if (!etlFieldExists(state, action.fieldName)) {
+          /* nothing to do */
+          return;
+        }
+
         if (isEtlFieldDerived(state, action.fieldName)) {
           next([
-            deleteDerivedField(action.fieldName),
-            tagWarehouseState('STALE'),
+            deleteDerivedField(action.fieldName), // document
+            tagWarehouseState('STALE'), // document
           ]);
         } else {
+          next(
+            setUiLoadingState({
+              toggle: true,
+              feature: ETL_VIEW,
+              message: 'Updating etlView - removing field',
+            }),
+          );
+
           const removeCfg = {
             /* eslint-disable no-shadow */
             hvsSelector: getHeaderViews,
             etlFieldChangesSelector: getEtlFieldChanges,
-            etlUnitsSelector: (state) => getEtlObject(state).etlUnits,
+            etlUnitsSelector: (s) => getEtlObject(s).etlUnits,
             DEBUG,
             /* eslint-enable no-shadow */
           };
 
+          // atomic state mutation
+          state = getState();
           const { headerViews } = removeNonDerivedEtlField(removeCfg)({
             fieldName: action.fieldName,
-            // purpose: action.purpose,
             state,
           });
-
-          next([
-            setHeaderViews(headerViews),
-            setEtlViewAsync(headerViews, getEtlFieldChanges(state), new Date()),
-          ]);
+          await setEtlViewAsync(next, state, {
+            headerViews,
+            startTime: new Date(),
+          });
         }
         break;
       }
 
       case RENAME_ETL_FIELD: {
         const { oldValue, newValue, etlFieldNameAndPurposeValues } = action;
-        const state = getState();
+
+        next(resetEtlViewErrors());
 
         const renameCfg = {
           hvsSelector: getHeaderViews,
@@ -194,9 +245,8 @@ const middleware =
           DEBUG,
         };
 
-        next(resetEtlViewErrors());
-
         try {
+          const state = getState();
           const { headerViews, etlFieldChanges } = renameEtlField(renameCfg)({
             oldValue,
             newValue,
@@ -206,14 +256,18 @@ const middleware =
           // when to call pivot(hvs) && applyFieldChanges(computedValue)
           if (typeof headerViews !== 'undefined') {
             next([
-              setHeaderViews(headerViews),
-              setEtlViewAsync(
-                headerViews,
-                etlFieldChanges || getEtlFieldChanges(state),
-                new Date(),
-              ),
+              setUiLoadingState({
+                toggle: true,
+                feature: ETL_VIEW,
+                message: 'Updating etlField: field name',
+              }),
               tagWarehouseState('STALE'),
             ]);
+
+            await setEtlViewAsync(next, state, {
+              headerViews,
+              startTime: new Date(),
+            });
           }
 
           // ...vs just call applyFieldChanges(etlObject)
@@ -225,12 +279,13 @@ const middleware =
               .applyFieldProps()
               .setGlobalSpanRef()
               .return();
+
             next([
               setNotification({
                 message: `Changed etlFieldChanges`,
                 feature: ETL_VIEW,
               }),
-              setEtlView(eo, ec),
+              setEtlViewAction(eo, ec),
               tagWarehouseState('STALE'),
             ]);
           } else {
@@ -257,17 +312,75 @@ const middleware =
 
 /**
  *
- * 🔖 This action gets augmented by async.middleware
+ * 🔖 This action gets processed by async.middleware
+ *
+ * 🔑 The headerViews state sometimes gets updated before this action
+ *    completes.  However, what we can't do, is use the "current" state
+ *    to compute the pivot value.
  *
  */
-function setEtlViewAsync(headerViews, etlFieldChanges, startTime) {
+async function setEtlViewAsync(dispatch, state, changes) {
+  // optional data
+  const {
+    startTime,
+    headerViews = getHeaderViews(state),
+    etlFieldChanges = getEtlFieldChanges(state),
+  } = changes;
+
+  // compute the value from state
+  const computedEtlObject = await pivot(headerViews);
+
+  // + user input and other
+  const final = await etlFromPivot()
+    .init(computedEtlObject, etlFieldChanges)
+    .removeStalePropChanges()
+    .removeStaleDerivedFields()
+    .addDerivedFields()
+    .applySourceSequences() // ⚠️  sequence matters (before applyFieldProps)
+    .applyFieldProps()
+    .setGlobalSpanRef()
+    .return();
+
+  const endTime = new Date();
+
+  // 🔖 do not use an array of dispatches
+  dispatch(setHeaderViewsAction(headerViews));
+  dispatch(
+    setEtlViewAction(
+      final.etlObject,
+      final.etlFieldChanges,
+      `${endTime - startTime} ms`,
+    ),
+  );
+  dispatch(
+    setUiLoadingState({
+      toggle: false,
+      feature: ETL_VIEW,
+      message: 'Done etlView async',
+    }),
+  );
+}
+export default middleware;
+
+/*
+function setEtlViewAsyncOld(data) {
   return {
-    type: 'GO LUCI',
-    meta: { feature: ETL_VIEW },
-    payload: (dispatch) => {
-      /* compute the value from state */
+    type: 'GO LUCI ASYNC',
+    meta: `${ETL_VIEW} ASYNC`,
+    payload: (dispatch, getState) => {
+      // latest state
+      const state = getState();
+      // optional data
+      const {
+        startTime,
+        headerViews = getHeaderViews(state),
+        etlFieldChanges = getEtlFieldChanges(state),
+      } = data;
+
+      // compute the value from state
       const computedEtlObject = pivot(headerViews);
 
+      // + user input and other
       const final = etlFromPivot()
         .init(computedEtlObject, etlFieldChanges)
         .removeStalePropChanges()
@@ -281,16 +394,103 @@ function setEtlViewAsync(headerViews, etlFieldChanges, startTime) {
       const endTime = new Date();
 
       // 🔖 do not use an array of dispatches
+      setHeaderViews(headerViews);
       dispatch(
-        setEtlView(
+        setEtlViewAction(
           final.etlObject,
           final.etlFieldChanges,
           `${endTime - startTime} ms`,
         ),
       );
-      dispatch(setLoader({ toggle: false, feature: ETL_VIEW }));
+      dispatch(setUiLoadingState({ toggle: false, feature: ETL_VIEW }));
     },
   };
-}
-
-export default middleware;
+} */
+/*
+const temp = {
+  __derivedFields: {
+    product: {
+      idx: 5,
+      name: 'product',
+      enabled: true,
+      purpose: 'mcomp',
+      levels: [
+        ['A', 45949],
+        ['C', 50764],
+      ],
+      'map-symbols': {
+        arrows: {},
+      },
+      'etl-unit': 'Rx',
+      format: null,
+      'null-value-expansion': null,
+      'map-files': {
+        arrows: {
+          '/shared/datafiles/8108ac51-8d8f-497a-91d3-6fef941655ed/dropbox/1db117/lE0WdposqDkAAAAAABpp3w/productComp_Units_mini.csv':
+            'A',
+          '/shared/datafiles/8108ac51-8d8f-497a-91d3-6fef941655ed/dropbox/1db117/lE0WdposqDkAAAAAABppzw/productA_Units.csv':
+            'C',
+        },
+      },
+      sources: [
+        {
+          enabled: true,
+          'map-implied': {
+            domain: 'NPI Number',
+            codomain: 'A',
+          },
+          'header-idx': null,
+          'field-alias': 'product',
+          'source-type': 'IMPLIED',
+          purpose: 'mcomp',
+          'null-value': null,
+          format: null,
+          'map-symbols': {
+            arrows: {},
+          },
+          levels: [['A', 45949]],
+          nlevels: 1,
+          filename:
+            '/shared/datafiles/8108ac51-8d8f-497a-91d3-6fef941655ed/dropbox/1db117/lE0WdposqDkAAAAAABpp3w/productComp_Units_mini.csv',
+          'map-weights': {
+            arrows: {
+              A: 1,
+            },
+          },
+        },
+        {
+          enabled: true,
+          'map-implied': {
+            domain: 'NPI Number',
+            codomain: 'C',
+          },
+          'header-idx': null,
+          'field-alias': 'product',
+          'source-type': 'IMPLIED',
+          purpose: 'mcomp',
+          'null-value': null,
+          format: null,
+          'map-symbols': {
+            arrows: {},
+          },
+          levels: [['C', 50764]],
+          nlevels: 1,
+          filename:
+            '/shared/datafiles/8108ac51-8d8f-497a-91d3-6fef941655ed/dropbox/1db117/lE0WdposqDkAAAAAABppzw/productA_Units.csv',
+          'map-weights': {
+            arrows: {
+              C: 1,
+            },
+          },
+        },
+      ],
+      'map-weights': {
+        arrows: {
+          A: 1,
+          C: 1,
+        },
+      },
+    },
+  },
+};
+*/
